+++ forms/form-optionals-category/index.tsx
@@ -33,11 +33,14 @@ import {
 	TableHeader,
 	TableRow,
 } from "@/components/ui/table";
+import { AlertDialogComponent } from "@/components/alert";
+import DialogType from "@/types/dialogs";
 
 export default function FormOptionalCategory() {
 	const [useLoading, setLoading] = useState(false);
 	const [useOptionalCategorys, setOptionalCategory] = useState<OptionalCategory[]>([]);
 	const [reload, serReload] = useState<boolean>(false);
+	const [propsRegister, setPropsRegister] = useState<DialogType>()
 
 	const { createOptionalCategory, getAllOptionalCategory, deleteOptionalCategory } = useOptionalCategory();
 	const { useSetFormOptionalCategory } = FormDataOptionalCategory();
@@ -94,25 +97,42 @@ export default function FormOptionalCategory() {
 	const handleDelete = (OptionalCategory: OptionalCategory) => {
 		if (!OptionalCategory) return;
 
-		deleteOptionalCategory(OptionalCategory.idOptionalCategory)
-			.then(() => {
-				serReload(!reload);
-				toast("Categoria de opcional deletada com sucesso!", {
-					description: new Date().toLocaleDateString("pt-BR"),
-					action: {
-						label: "Fechar",
-						onClick: () => console.log("Undo"),
-					},
-				})
-			}).catch((error) => {
-				toast.error("Erro ao deletar categoria opcional:", {
-					description: error,
-					action: {
-						label: "Fechar",
-						onClick: () => console.log("Undo"),
-					},
-				});
-			});
+		setPropsRegister({
+			title: 'Deseja excluir?',
+			description: `A Categoria ${OptionalCategory.description} pode estar sendo usada em outros veículos e gerar erros!`,
+			confirm: () => {
+
+				deleteOptionalCategory(OptionalCategory.idOptionalCategory)
+					.then(() => {
+						serReload(!reload);
+						toast("Categoria de opcional deletada com sucesso!", {
+							description: new Date().toLocaleDateString("pt-BR"),
+							action: {
+								label: "Fechar",
+								onClick: () => console.log("Undo"),
+							},
+						})
+						setPropsRegister({ ...propsRegister, open: false })
+
+					}).catch((error) => {
+						toast.error("Erro ao deletar categoria opcional:", {
+							description: error,
+							action: {
+								label: "Fechar",
+								onClick: () => console.log("Undo"),
+							},
+						});
+					});
+			},
+			cancel: () => setPropsRegister({ ...propsRegister, open: false }),
+			cancelText: 'cancelar',
+			confirmText: 'excluir',
+			cancelButton: true,
+			confirmButton: true,
+			open: true
+		})
+
+
 	}
 
 	const MakeTableRow = () => {
@@ -174,6 +194,8 @@ export default function FormOptionalCategory() {
 
 	return <Form {...useSetFormOptionalCategory}>
 		<Toaster />
+		<AlertDialogComponent {...propsRegister} />
+
 		<div className="w-full max-w-7xl mx-auto px-4 py-8 space-y-6">
 			<div className="text-center">
 				<h2 className="text-3xl font-bold tracking-tight text-foreground sm:text-4xl">
